[what][style] 处理 msvc 2022 编译警告

+++ H26xBinaryReader.cpp
@@ -72,7 +72,7 @@ void H26xBinaryReader::UE(uint32_t& value)
         assert(b == 0 || b == 1);
     }
     U(leadingZeroBits, tmp);
-    value = (1 << leadingZeroBits) - 1 + tmp;
+    value = (uint32_t)((1 << leadingZeroBits) - 1 + tmp);
 }
 
 void H26xBinaryReader::SE(int32_t& value)
@@ -95,8 +95,8 @@ void H26xBinaryReader::SE(int32_t& value)
                                                 do\
                                                 {\
                                                     ReadOneByteAuto();\
-                                                    size_t readBits = bits <= (8 - (size_t)_curBitPos)? bits : (8 - (size_t)_curBitPos);\
-                                                    bits = bits - readBits;\
+                                                    size_t readBits = (size_t)bits <= (8 - (size_t)_curBitPos)? bits : (8 - (size_t)_curBitPos);\
+                                                    bits = (size_t)(bits - readBits);\
                                                     value <<= readBits;\
                                                     if (readBits < 8 && !firstFlag)\
                                                     {\
@@ -228,7 +228,7 @@ void H26xBinaryReader::Skip(size_t bits)
 {
     if (bits + _curBitPos < 8) // 不需要跳转至下一个字节
     {
-        _curBitPos = bits + _curBitPos;
+        _curBitPos = uint8_t(bits + _curBitPos);
         return;
     }
     else // 需要跳转的 bits 先用消费当前 byte 剩余的 bits
